backend/requirements.txt
asgiref==3.10.0
Django==5.2.8
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==10.2.0
PyJWT==2.10.1
python-dotenv==1.0.1
sqlparse==0.5.3
